backend/requirements.txt
annotated-types==0.7.0
blinker==1.9.0
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
docstring_parser==0.16
Faker==37.0.2
firebase-admin==6.7.0
Flask==3.1.0
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.30.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
msgpack==1.1.0
numpy==2.2.4
packaging==24.2
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
shapely==2.0.7
six==1.17.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vertexai==1.71.1
Werkzeug==3.1.3
# Google Cloud dependencies
google-cloud-speech>=2.23.0
google-cloud-texttospeech>=2.14.1
google-cloud-aiplatform>=1.36.3
google-cloud-language>=2.11.1
google-cloud-storage>=2.10.0

# Firebase dependencies
firebase-admin>=6.2.0

# Web API dependencies
flask>=2.3.3
flask-cors>=4.0.0
gunicorn>=21.2.0

# Other dependencies
numpy>=1.24.0
python-dotenv>=1.0.0